# Services/opsmind-ticket-service/swagger.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "OpsMind Ticket Service API",
                 "version":  "1.0.0",
                 "description":  "REST API for managing tickets."
             },
    "servers":  [
                    {
                        "url":  "/"
                    }
                ],
...
    "paths":  {
                  "/tickets":  {
                                   "post":  {
                                                "tags":  [
                                                             "Tickets"
                                                         ],
                                                "summary":  "Create a ticket",
                                                "description":  "User-provided fields: title, description, type_of_request, requester_id, latitude, longitude. Priority, support level, and initial status (OPEN) are system-assigned. GPS coordinates are used by the Workflow Service for location-aware technician assignment weighted by proximity, workload, and priority.",
                                                "requestBody":  {
                                                                    "required":  true,
                                                                    "content":  {
                                                                                    "application/json":  {
                                                                                                             "schema":  {
                                                                                                                            "type":  "object",
                                                                                                                            "required":  [
                                                                                                                                             "title",
                                                                                                                                             "description",
                                                                                                                                             "type_of_request",
                                                                                                                                             "requester_id",
                                                                                                                                             "latitude",
                                                                                                                                             "longitude"
                                                                                                                                         ],
                                                                                                                            "properties":  {
                                                                                                                                               "title":  {
                                                                                                                                                             "type":  "string"
                                                                                                                                                         },
                                                                                                                                               "description":  {
                                                                                                                                                                   "type":  "string"
                                                                                                                                                               },
                                                                                                                                               "type_of_request":  {
                                                                                                                                                                       "type":  "string",
                                                                                                                                                                       "enum":  [
                                                                                                                                                                                    "INCIDENT",
                                                                                                                                                                                    "SERVICE_REQUEST",
                                                                                                                                                                                    "MAINTENANCE"
                                                                                                                                                                                ]
                                                                                                                                                                   },
                                                                                                                                               "requester_id":  {
                                                                                                                                                                    "type":  "string",
                                                                                                                                                                    "description":  "UUID of the user submitting the ticket"
                                                                                                                                                                },
                                                                                                                                               "latitude":  {
                                                                                                                                                                "type":  "number",
                                                                                                                                                                "minimum":  -90,
                                                                                                                                                                "maximum":  90,
                                                                                                                                                                "description":  "GPS latitude of the incident location — used for intelligent assignment"
                                                                                                                                                            },
                                                                                                                                               "longitude":  {
                                                                                                                                                                 "type":  "number",
                                                                                                                                                                 "minimum":  -180,
                                                                                                                                                                 "maximum":  180,
                                                                                                                                                                 "description":  "GPS longitude of the incident location — used for intelligent assignment"
                                                                                                                                                             }
                                                                                                                                           }
                                                                                                                        }
                                                                                                         }
                                                                                }
                                                                },
                                                "responses":  {
                                                                  "201":  {
                                                                              "description":  "Created"
                                                                          }
                                                              }
                                            },
                                   "get":  {
                                               "tags":  [
                                                            "Tickets"
                                                        ],
                                               "summary":  "List tickets",
                                               "description":  "Returns all tickets except soft-deleted (is_deleted = false).",
                                               "parameters":  [
                                                                  {
                                                                      "in":  "query",
                                                                      "name":  "status",
                                                                      "schema":  {
                                                                                     "type":  "string",
                                                                                     "enum":  [
                                                                                                  "OPEN",
                                                                                                  "IN_PROGRESS",
                                                                                                  "RESOLVED",
                                                                                                  "CLOSED"
                                                                                              ]
                                                                                 }
                                                                  },
                                                                  {
                                                                      "in":  "query",
                                                                      "name":  "priority",
                                                                      "schema":  {
                                                                                     "type":  "string",
                                                                                     "enum":  [
                                                                                                  "LOW",
                                                                                                  "MEDIUM",
                                                                                                  "HIGH"
                                                                                              ]
                                                                                 }
                                                                  },
                                                                  {
                                                                      "in":  "query",
                                                                      "name":  "requester_id",
                                                                      "schema":  {
                                                                                     "type":  "string"
                                                                                 }
                                                                  },
                                                                  {
                                                                      "in":  "query",
                                                                      "name":  "limit",
                                                                      "schema":  {
                                                                                     "type":  "integer",
                                                                                     "default":  50
                                                                                 }
                                                                  },
                                                                  {
                                                                      "in":  "query",
                                                                      "name":  "offset",
                                                                      "schema":  {
                                                                                     "type":  "integer",
                                                                                     "default":  0
                                                                                 }
                                                                  }
                                                              ],
                                               "responses":  {
                                                                 "200":  {
                                                                             "description":  "OK"
                                                                         }
                                                             }
                                           }
                               },
...
              },
...
}